+++ spec/ofx_kit/configuration_spec.rb
@@ -7,22 +7,22 @@ RSpec.describe OFX::Configuration do
   subject(:config) { described_class.new }
 
   describe 'default mappings' do
-    it 'loads transaction mappings from YAML' do
+    it 'loads transaction mappings' do
       expect(config.xml_mappings_for(:transaction)['FITID']).to eq('fit_id')
       expect(config.xml_mappings_for(:transaction)['TRNAMT']).to eq('amount')
       expect(config.xml_mappings_for(:transaction)['MEMO']).to eq('memo')
     end
 
-    it 'loads bank_account mappings from YAML' do
+    it 'loads bank_account mappings' do
       expect(config.xml_mappings_for(:bank_account)['BANKID']).to eq('bank_id')
       expect(config.xml_mappings_for(:bank_account)['ACCTID']).to eq('account_id')
     end
 
-    it 'loads credit_card_account mappings from YAML' do
+    it 'loads credit_card_account mappings' do
       expect(config.xml_mappings_for(:credit_card_account)['ACCTID']).to eq('account_id')
     end
 
-    it 'loads balance mappings from YAML' do
+    it 'loads balance mappings' do
       expect(config.xml_mappings_for(:balance)['BALAMT']).to eq('amount')
     end
 
@@ -63,90 +63,6 @@ RSpec.describe OFX::Configuration do
     end
   end
 
-  describe 'auto-load from project path' do
-    let(:auto_load_path) { File.join(__dir__, '../fixtures/auto_mappings.yml') }
-
-    after { File.delete(auto_load_path) if File.exist?(auto_load_path) }
-
-    it 'loads the project file on top of gem defaults when it exists' do
-      File.write(auto_load_path, "FIELDS:\n  STMTTRN:\n    CUSTFIELD: \"auto_attr\"\n")
-      config = described_class.new(auto_load_path: auto_load_path)
-      expect(config.xml_mappings_for(:transaction)['CUSTFIELD']).to eq('auto_attr')
-      expect(config.xml_mappings_for(:transaction)['FITID']).to eq('fit_id')
-    end
-
-    it 'skips auto-load when the project file does not exist' do
-      config = described_class.new(auto_load_path: '/nonexistent/ofx_mappings.yml')
-      expect(config.xml_mappings_for(:transaction)['FITID']).to eq('fit_id')
-    end
-
-    it 'raises ConfigurationError when auto-loaded file tries to override a core field' do
-      File.write(auto_load_path, "FIELDS:\n  STMTTRN:\n    TRNAMT: \"foobar\"\n")
-      expect { described_class.new(auto_load_path: auto_load_path) }
-        .to raise_error(OFX::ConfigurationError, /Cannot override core mapping.*TRNAMT/)
-    end
-  end
-
-  describe '#load_mappings' do
-    let(:custom_yaml_path) { File.join(__dir__, '../fixtures/custom_mappings.yml') }
-
-    before do
-      File.write(custom_yaml_path, <<~YAML)
-        FIELDS:
-          STMTTRN:
-            CUSTFIELD: "custom_attr"
-      YAML
-    end
-
-    after { File.delete(custom_yaml_path) if File.exist?(custom_yaml_path) }
-
-    it 'merges custom YAML on top of defaults' do
-      config.load_mappings(custom_yaml_path)
-      expect(config.xml_mappings_for(:transaction)['CUSTFIELD']).to eq('custom_attr')
-      expect(config.xml_mappings_for(:transaction)['FITID']).to eq('fit_id')
-    end
-
-    it 'raises ConfigurationError for a missing file' do
-      expect { config.load_mappings('nonexistent.yml') }
-        .to raise_error(OFX::ConfigurationError, /not found/)
-    end
-
-    it 'raises ConfigurationError when FIELDS key is missing' do
-      File.write(custom_yaml_path, "STMTTRN:\n  FOO: bar\n")
-      expect { config.load_mappings(custom_yaml_path) }
-        .to raise_error(OFX::ConfigurationError, /missing.*FIELDS/)
-    end
-
-    it 'does not raise when FIELDS is present and OPTIONS is absent' do
-      File.write(custom_yaml_path, "FIELDS:\n  STMTTRN:\n    CUSTFIELD: custom_attr\n")
-      expect { config.load_mappings(custom_yaml_path) }.not_to raise_error
-    end
-
-    it 'raises ConfigurationError for an unknown XML tag' do
-      File.write(custom_yaml_path, "FIELDS:\n  UNKNOWN_TAG:\n    FOO: bar\n")
-      expect { config.load_mappings(custom_yaml_path) }
-        .to raise_error(OFX::ConfigurationError, /Unknown section/)
-    end
-
-    it 'raises ConfigurationError when a mapping value is not a Hash' do
-      File.write(custom_yaml_path, "FIELDS:\n  STMTTRN: \"oops this should be a hash\"\n")
-      expect { config.load_mappings(custom_yaml_path) }
-        .to raise_error(OFX::ConfigurationError, /must be a Hash/)
-    end
-
-    it 'raises ConfigurationError when overriding a core field via load_mappings' do
-      File.write(custom_yaml_path, "FIELDS:\n  STMTTRN:\n    TRNAMT: \"foobar\"\n")
-      expect { config.load_mappings(custom_yaml_path) }
-        .to raise_error(OFX::ConfigurationError, /Cannot override core mapping.*TRNAMT/)
-    end
-
-    it 'raises ConfigurationError when overriding CURDEF via load_mappings' do
-      File.write(custom_yaml_path, "FIELDS:\n  STMTRS:\n    CURDEF: \"foobar\"\n")
-      expect { config.load_mappings(custom_yaml_path) }
-        .to raise_error(OFX::ConfigurationError, /Cannot override core mapping.*CURDEF/)
-    end
-  end
-
   describe 'multi_statement_warnings' do
     it 'defaults to true' do
       expect(config.multi_statement_warnings?).to be true
@@ -157,5 +73,4 @@ RSpec.describe OFX::Configuration do
       expect(config.multi_statement_warnings?).to be false
     end
   end
-
 end